Move getFiltersFromUrl to separate file for easier testing
